% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_nzffd.R
\name{clean_nzffd}
\alias{clean_nzffd}
\title{Clean NZ freshwater Fish fishdasets}
\usage{
clean_nzffd(fishd)
}
\arguments{
\item{fishd}{A dataframe imported from the NZFFD using \code{import_nzffd}}
}
\value{
A dataframe of the same dimensions as \code{fishd}, but cleaned.
}
\description{
Clean up data imported from the NZ Freshwater Fish Database
}
\details{
If I want a details section this is where I should write it.
}
\examples{
\dontrun{

dat <- clean_nzffd(nzffdr::nzffd_data)
head(dat)
}
}
