% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_nzffd.R
\name{coord_nzffd}
\alias{coord_nzffd}
\title{NZMG coordinate conversion}
\usage{
coord_nzffd(east, north, projargs = "+init=epsg:4326")
}
\arguments{
\item{east}{A vector of New Zealand Map Grid eastings}

\item{north}{A vector of New Zealand Map Grid northings}
}
\value{
a dataframe of two columns, the new eastings and northings#'
}
\description{
update New Zealand Map Grid coordinates (the default in the freshwater fish 
database) to another projection.
}
\details{
This function is a wrapper for a series of functions from the 
sp package.

REC citation:
}
\examples{
\dontrun{

df <- coords_nzffd(east = nzffdr::nzffd_data$east, 
                   north = nzffdr::nzffd_data$north)
head(df)
}
}
