% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_nzffd.R
\name{fill_nzffd}
\alias{fill_nzffd}
\title{Add missing data to the NZFFD}
\usage{
fill_nzffd(fishd, alt = TRUE, maps = TRUE)
}
\arguments{
\item{fishd}{A dataframe imported from the NZFFD using \code{import_nzffd}.}

\item{alt}{Should missing elevation data be added to the dataframe.}

\item{maps}{Should missing map tiles be added to the dataframe.}
}
\value{
A dataframe with the same number of rows as \code{fishd} but with
four additional columns added (\code{family, sci_name, common_name, threat_status}).
}
\description{
Adds family, genus, species and common names, the species threat
classification status and missing values for the \code{map} and \code{alt}
variables.
}
\details{
Altitude data is added by taking the site coordinates and pulling the relevant
data from an 8m DEM (url). Note values taken from the 8m DEM are not 100
percent accurate and should be treated with caution (*add MoE??*).

If \code{alt = TRUE} this function requires an internet connection to query
the 8m DEM.
}
\examples{
\dontrun{

df <- fill_nzffd(nzffdr::nzffd_data)
head(df)
}
}
