% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzffdr_old.R
\name{nzffd_add}
\alias{nzffd_add}
\title{Add missing data (depreciated)}
\usage{
nzffd_add(...)
}
\arguments{
\item{...}{arguments to pass to function.}
}
\value{
A dataframe where each row is a NZFFD record.
}
\description{
Add in missing NZFFD data. This function is now depreciated. See the \href{https://flee598.github.io/nzffdr/}{help page} for details.
}
\examples{
\dontrun{
nzffd_add()
}
}
