% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzffdr_old.R
\name{nzffd_fill}
\alias{nzffd_fill}
\title{Fill missing data (depreciated)}
\usage{
nzffd_fill(...)
}
\arguments{
\item{...}{arguments to pass to function.}
}
\value{
A dataframe where each row is a NZFFD record.
}
\description{
Fill in missing NZFFD data. This function is now depreciated. See the \href{https://flee598.github.io/nzffdr/}{help page} for details.
}
\examples{
\dontrun{
nzffd_fill()
}
}
