% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nzffd_method}
\alias{nzffd_method}
\title{Dataframe listing all fishing methods used in the NZFFD}
\format{
A dataframe of 59 rows and one variable:
\describe{
  \item{method}{fishing method used}
}
}
\usage{
nzffd_method
}
\description{
Dataframe listing the possible inputs for the fish_method argument
of the nzffd_import function.
}
\keyword{datasets}
