% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzffdr_get_table.R
\name{nzffdr_get_table}
\alias{nzffdr_get_table}
\title{Get NZFFD search terms}
\usage{
nzffdr_get_table(x = c("fish_method", "institution", "taxon"))
}
\arguments{
\item{x}{one of "fish_method", "institution" or "taxon", depending on
which argument values are wanted.}
}
\value{
A dataframe or character string of argument options.
}
\description{
Lists of possible argument options for function \code{nzffdr_import()}.
}
\details{
Returns the possible argument values for arguments: fish_method, 
institution and taxon, used in the function \code{nzffdr_import()}.

This function requires an internet connection to query NIWA's database.
}
\examples{
\dontrun{

dat <- nzffdr_get_table("taxon")

}
}
