% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzffdr_ind_lengths.R
\name{nzffdr_ind_lengths}
\alias{nzffdr_ind_lengths}
\title{Fish length to tidy long format}
\usage{
nzffdr_ind_lengths(fishd)
}
\arguments{
\item{fishd}{an NZFFD dataframe returned from [nzffdr_import()]. Must contain
the columns "nzffdRecordNumber", "taxonName" and "indLengths".}
}
\value{
A dataframe with three columns,"nzffdRecordNumber", "taxonName" and
 "indLengths".
}
\description{
Converts individual fish length measures from multiple entries in
a single cell to tidy long format.
}
\examples{
nzffdr_ind_lengths(nzffdr::nzffdr_data)
}
