% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nzffdr_nzmap}
\alias{nzffdr_nzmap}
\title{Simple features map of New Zealand}
\format{
A simple features dataframe with 4 rows and 2 columns:
\describe{
  \item{island}{Island name}
  \item{geometry}{Line geometry}
}
}
\source{
\url{https://data.linz.govt.nz}
}
\usage{
nzffdr_nzmap
}
\description{
A simple features map of New Zealand. A simplified version of the 1:150k
NZ map outline available from Land Information New Zealand.
CRS: NZ Transverse Mercator (NZTM: EPSG 2193).
}
\keyword{datasets}
