% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzffdr_razzle_dazzle.R
\name{nzffdr_razzle_dazzle}
\alias{nzffdr_razzle_dazzle}
\title{Wrapper for multiple nzffdr functions}
\usage{
nzffdr_razzle_dazzle()
}
\value{
An NZFFD dataframe which has been cleaned and had date, taxonomic and 
threat classification status data added.
}
\description{
Wraps multiple nzffdr functions allowing, importing, cleaning and adding
of new information all in one step.
}
\details{
Wraps: [nzffdr_import()], [nzffdr_clean()], [nzffdr_add_date()],
[nzffdr_taxon_threat()], [nzffdr_widen_habitat()] and runs the lot
in one go, returning a downloaded and cleaned NZFFD dataset.

This function requires an internet connection to query the NZFFD
}
\examples{
\dontrun{
dat <- nzffdr_razzle_dazzle()
}
}
