% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_nzffd}
\alias{species_nzffd}
\title{Dataframe listing all species in the NZFFD}
\format{
A dataframe of 75 rows and two variables:
\describe{
  \item{sci}{Genus and species}
  \item{common}{common name}
}
}
\usage{
species_nzffd
}
\description{
A dataframe listing the possible inputs for the species argument
of the import_nzffd function. Either the scientific or common name can
be used.
}
\keyword{datasets}
