% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nzffd.R
\name{add_nzffd}
\alias{add_nzffd}
\title{Add River Environment Classification (REC) data to the NZFFD}
\usage{
add_nzffd(fishd)
}
\arguments{
\item{fishd}{A dataframe imported from the NZFFD using \code{import_nzffd}.}
}
\value{
A dataframe with the same number of rows as \code{fishd} but
with *XX* additional columns from the REC database added.
}
\description{
Use the reachID variable to match NZFFD sample locations against the REC database and add
coresponding REC data to the NZFFD.
}
\details{
Note that any row that has an \code{nzreach} value of \code{0}, \code{9}
or \code{NA} will not return any REC data.

This function requires an internet connection to query the REC database.

REC citation:
}
\examples{
\dontrun{

df <- add_nzffd(nzffdr::nzffd_data)
head(df)
}
}
