% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{method_nzffd}
\alias{method_nzffd}
\title{Dataframe listing all fishing methods used in the NZFFD}
\format{
A dataframe of 59 rows and one variable:
\describe{
  \item{method}{fishing method used}
}
}
\usage{
method_nzffd
}
\description{
Dataframe listing the possible inputs for the fish_method argument
of the import_nzffd function.
}
\keyword{datasets}
