% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nzffd_nzmap}
\alias{nzffd_nzmap}
\title{Simple features map of New Zealand}
\format{
A simple features dataframe with 4 rows and 3 columns:
\describe{
  \item{name}{Island name}
  \item{name_ascii}{Island name ascii characters}
  \item{geometry}{Line geometry}
}
}
\source{
\url{https://data.linz.govt.nz}
}
\usage{
nzffd_nzmap
}
\description{
A simple features map of New Zealand. A simplified version of the 1:150k
NZ map outline available from Land Information New Zealand.
CRS: New Zealand Map Grid, EPSG:27200 (matching NZFFD observations). 
The map contains polygons for the three main island plus the Chatham Islands.
}
\keyword{datasets}
