% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nzffdr_data}
\alias{nzffdr_data}
\title{Sample NZFFD data.}
\format{
A dataframe of 200 rows and 67 variables.
}
\source{
\href{https://niwa.co.nz}{NIWA}
}
\usage{
data(nzffdr_data)
}
\description{
Sample NZFFD data.
}
\keyword{datasets}
